Web API Test 9.19

+++ Bassa/Inventry Management System - ITP/BackEnd/DataAccess/Persistence/Repositories/Repository.cs
@@ -19,7 +19,7 @@ namespace DataAccess.Persistence.Repositories
         }
 
         //Get Methods
-        public TEntity Get(int Id)
+        public TEntity Get(long Id)
         {
             return this._entity.Find(Id);
         }
